feat: Update URL from search
this uses a custom extension, which allows to push paremeters to arbitrary URLs,
in order to push query parameters without the original path.

This is handy in order to build sharable URLs dynamically.

One current drawback is the lack of sorting, which I handled in rust before.
I will resolve this by prerendering the search form with two hidden parameters,
which are prerendered in rust later on.

For now having the search destroy the previous ordering is fine.

The extension has been provided by Daniel Hernandez in
https://github.com/bigskysoftware/htmx/issues/2156#issuecomment-1968042842

Co-authored-by: Daniel Hernandez <daniel95.hg@gmail.com>